+++ src/operations.h
@@ -79,7 +79,7 @@ namespace sharp {
   /*
     Trim an image
   */
-  VImage Trim(VImage image, std::vector<double> background, double const threshold);
+  VImage Trim(VImage image, std::vector<double> background, double threshold, bool const lineArt);
 
   /*
    * Linear adjustment (a * in + b)